Added DMA generics
