+++ Autopilot/Path Manager/PathManager.c
@@ -15,7 +15,8 @@
 #include "StartupErrorCodes.h"
 
 #include "../Common/Interfaces/SPI.h"
-#include "InterchipDMA.h"
+#include "../Common/Utilities/InterchipDMA.h"
+#include "GPSDMA.h"
 
 #if DEBUG
 #include <stdio.h>
@@ -57,8 +58,8 @@ void pathManagerInit(void) {
 
 
 //Communication with GPS
-    initSPI(2, 0, SPI_MODE1, SPI_WORD, SPI_SLAVE);
     init_DMA2();
+    initSPI(2, 0, SPI_MODE1, SPI_WORD, SPI_SLAVE);
     initMainBatterySensor();
     initAirspeedSensor();
 
@@ -72,12 +73,13 @@ void pathManagerInit(void) {
     TRISBbits.TRISB4 = 1;   //Init RB4 as Input (1)
     TRISBbits.TRISB5 = 1;   //Init RB5 as Input (1)
 
-    INTERCOM_3 = 0;    //Set RA12 to Output a Value of 0
-    INTERCOM_4 = 0;    //Set RA13 to Output a Value of 0
+//    INTERCOM_3 = 0;    //Set RA12 to Output a Value of 0
+//    INTERCOM_4 = 0;    //Set RA13 to Output a Value of 0
 
+    init_DMA0(0);
+    init_DMA1(0);
     initSPI(1, 0, SPI_MODE1, SPI_BYTE, SPI_MASTER);
-    init_DMA0();
-    init_DMA1();
+
     DMA1REQbits.FORCE = 1;
     while (DMA1REQbits.FORCE == 1);
