+++ Autopilot/Path Manager/nbproject/configurations.xml
@@ -12,6 +12,7 @@
       <logicalFolder name="f1" displayName="Utilities" projectFiles="true">
         <itemPath>../Common/Utilities/Logger.h</itemPath>
         <itemPath>../Common/Utilities/ByteQueue.h</itemPath>
+        <itemPath>../Common/Utilities/InterchipDMA.h</itemPath>
       </logicalFolder>
       <itemPath>Dubins.h</itemPath>
       <itemPath>InterchipDMA.h</itemPath>
@@ -40,6 +41,7 @@
       <logicalFolder name="f2" displayName="Utilities" projectFiles="true">
         <itemPath>../Common/Utilities/ByteQueue.c</itemPath>
         <itemPath>../Common/Utilities/Logger.c</itemPath>
+        <itemPath>../Common/Utilities/InterchipDMA.c</itemPath>
       </logicalFolder>
       <itemPath>Dubins.c</itemPath>
       <itemPath>InterchipDMA.c</itemPath>
